Folder cleanup and removed unused packages
