--- .eslintrc.json
@@ -1,38 +1,0 @@
-{
-    "env": {
-        "browser": true,
-        "es2021": true
-    },
-    "extends": [
-        "plugin:react/recommended",
-        "airbnb"
-    ],
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-        "ecmaFeatures": {
-            "jsx": true
-        },
-        "ecmaVersion": "latest",
-        "sourceType": "module"
-    },
-    "plugins": [
-        "react",
-        "@typescript-eslint"
-    ],
-    "rules": {
-        "react/react-in-jsx-scope": 0,
-        "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx", ".ts", ".tsx"] }],
-        "import/prefer-default-export": 0,
-        "import/extensions": 0,
-        "import/no-unresolved": 0,
-        "no-unused-vars": 0,
-        "react/jsx-props-no-spreading": 0,
-        "react/jsx-no-constructed-context-values": 0,
-        "jsx-a11y/anchor-is-valid": 0,
-        "arrow-body-style": 0,
-        "react/function-component-definition": 0,
-        "react/prop-types": 0,
-        "react/require-default-props": 0,
-        "no-nested-ternary": 0
-    }
-}
