+++ package.json
@@ -12,7 +12,6 @@
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.9.0",
     "date-fns": "^2.29.3",
-    "framer-motion": "^6.3.0",
     "hyperspace-client-js": "^1.3.8",
     "next": "latest",
     "react": "^18.2.0",
@@ -25,12 +24,6 @@
     "@types/react-dom": "^18.0.0",
     "@typescript-eslint/eslint-plugin": "^5.38.0",
     "@typescript-eslint/parser": "^5.38.0",
-    "eslint": "^8.23.1",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.6.1",
-    "eslint-plugin-react": "^7.31.8",
-    "eslint-plugin-react-hooks": "^4.6.0",
     "typescript": "^4.7.2"
   }
 }
