+++ src/pages/_document.tsx
@@ -10,7 +10,6 @@ export default class Document extends NextDocument {
       <Html>
         <Head />
         <body>
-          {/* Make Color mode to persists when you refresh the page. */}
           <ColorModeScript initialColorMode={theme.config.initialColorMode} />
           <Main />
           <NextScript />
