--- src/pages/project/Tokens.tsx
@@ -1,12 +1,0 @@
-import { Box, Image } from '@chakra-ui/react';
-import { FC } from 'react';
-
-type TokensProps = {}
-
-const Tokens: FC<TokensProps> = ({ projectData }) => {
-  return (
-   
-  );
-};
-
-export default Tokens;
